Aligns CMS section with draft-ietf-lamps-cms-ml-dsa-02.
This commit fixes the handling of the hashing when using composites
within CMS so it aligns with draft-ietf-lamps-cms-ml-dsa-02.
Since the referenced draft now only requires SHA-512, the table with
signature algorithm to hash mapping is not needed anymore.
Also the input to Composite-ML-DSA.Sign() is now the unhashed content
in case signed attributes are absent.

Closes #110