Fix RSA inconsistencies
All RSA3072 variants were using SHA-256 as the RSA signature's hash
function, except id-MLDSA87-RSA3072-PSS-SHA512. Made this consistent.

Elsewhere, the RSA3072-PSS MGF hash function was specified as SHA-512,
but it is the RSA signatures's hash function, so make it SHA-256.

SHA-256 was used because RSA3072 provides ~128 bits of security.

RSA4096-PSS wasn't consistent about the RSA hash function, sometimes
it said SHA-384, sometimes it said SHA-512. Set it to SHA-384 because
RSA4096 gives about 140 bits of security. SHA-512 could alternatively
be used, but it has to be consistent.

Refactor the test vector generation code so that the parameters
for a signature component are all defined in one location. And
align these parameters with the draft.

Regenrated text vectors based on the corrected parameters.

Fixes #242